Sources/backend-api/requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
Flask==3.1.0
Flask-Cors==5.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
numpy==2.2.4
pillow==11.1.0
pycparser==2.22
python-dotenv==1.0.1
reedsolo==1.7.0
scipy==1.15.2
Werkzeug==3.1.3
